chore: move entity-specific Settings pages to their directories
