+++ src/components/pages/TestSuites/TestSuiteDetails/TestSuiteSettings/SettingsTests/Nodes/IntersectionNode.tsx
